chore: restructure notification tests for improved clarity and organization
Split notification test cases into more focused and dedicated test classes. Renamed test files and classes for better consistency and alignment with their purposes.
